- Added command to edit landblock encounters.
- CustomDM: rebalanced Palenqual's Cavern, Hunter's Leap and the area right outside Palenqual's Cavern to be more level appropriate.

- CustomDM: Fixed extra xp from magic casters no being awarded.

+++ Source/ACE.Server/WorldObjects/Creature_Death.cs
@@ -257,7 +257,7 @@ namespace ACE.Server.WorldObjects
 
             double hitpointsXp = hitpoints / 10 * baseXp / 35;
 
-            double casterXp = baseXp * (numSpellInSpellbook / 20);
+            double casterXp = baseXp * ((float)numSpellInSpellbook / 20);
 
             return (int)Math.Round(baseXp + hitpointsXp + casterXp);
         }
